Downloads/Ai-Avengers-main/Ai-Avengers-main/backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
reportlab
python-multipart
